Add --random option to other flags
When using the --cc, --sc, --p2p, and --tor flags, one can
use either the --fastest or --random flag to specify to
connect to the fastest server or a random server.
For example, `protonvpn --sc --fastest` will connect to
the fastest secure core server,
and `protonvpn --sc --random` will connect to a random
secure core server.
If neither the --fastest nor the --random flags are used,
it will default to the fastest server.